requirements.txt
streamlit
torch
nltk
psycopg2-binary
scikit-learn
numpy